requirement.txt
langchain-community==0.2.1
langchain-core==0.2.1
crewai==0.28.8
litellm==1.40.52
langchain-ollama==0.1.2
streamlit==1.36.0
python-docx==1.1.2